requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
numpy==2.1.1
regex==2024.9.11
tqdm==4.66.5
rapidfuzz==3.9.7
httpx==0.27.2
tenacity==8.5.0
streamlit==1.38.0
markdown>=3.6